refactor: move whole game logic to one hook, delete useScore hook
